back/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests>=2.28.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
paho-mqtt>=2.0.0
pymongo>=4.0.0
qrcode[pil]>=7.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.0.0
joblib>=1.3.0
prophet>=1.1.0
gunicorn
pywebpush>=2.0.0

